refactor: Use direct sync_once response
I've just wasted most of the day trying to get the event under the API
to do what I want it to do. I finally realised that it's not actually
capable of that and instead I need to use the sync responses directly,
both for this and possibly for the room update triggers (i'll have to
look into that again later). However, I can now finally get the room
information that I need - haidily I did those sync changes yesterday,
which gives me easy access to the sync data and a way to retry without
losing any messages if something I do fails for the sync. I'll need to
make sure that the whole thing is transactional, though. Hey, progress!