fix: Correctly handle state in the timeline
The code change here adds support for the state after extension - and
properly handles the case where state is before the data. Before, there
would be a portion of state updates that we wouldn't see / handle.

This also updates all our deps, and removes the sqlx fork - the latest
alpha has an up-to-date sqlite version.